requirements.txt

streamlit==1.37.1
pillow==10.4.0
numpy==1.26.4
scikit-image==0.22.0
opencv-python==4.10.0.84
imagehash==4.3.1
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
pydantic==2.9.2
python-dotenv==1.0.1
reportlab==4.2.2
matplotlib==3.9.0
scikit-learn==1.4.2
datasets==3.0.1
